_getDataType numerics should be floats, not integers
git-svn-id: http://mytardis.googlecode.com/svn/branches/single-search-box-shauno@1585 1bbb43ac-ffe0-11de-8c0c-f1c8ac664a8e

+++ tardis/tardis_portal/search_indexes.py
@@ -11,7 +11,7 @@ from models import ParameterName
 
 def _getDataType(param_name):
     if param_name.isNumeric():
-        return IntegerField()
+        return FloatField()
     elif param_name.isDateTime():
         return DateTimeField()
     else:
